+++ g8/main.go
@@ -29,7 +29,7 @@ import (
 
 const (
 	// Version of go-giter8
-	Version = "0.4.0"
+	Version = "0.4.1"
 )
 
 func main() {