+++ RELEASE-NOTES.md
@@ -1,5 +1,11 @@
 # go-giter8 release notes
 
+## 2020-01-20 - v0.4.1
+
+- Fix: filename of file in "verbatim" list is kept intact.
+- Fix: `text/template`'s default delims (`{{` and `}}`) may cause issue sometimes.
+
+
 ## 2020-01-01 - v0.4.0
 
 - Support scaffolding & `scaffold` command.